runtime/vm: split constant folding into constfold.go
Pure code motion: constFold and its evaluators
(evalUnaryConst/evalBinaryConst/evalIndexConst/evalSliceConst,
toList, isZeroValue, isOneValue) move from liveness.go to a new
constfold.go. liveness.go retains the dataflow analysis (Liveness,
useDef, defRegs) plus the structural pass entry point (Optimize,
removeUnreachable).

liveness.go drops from ~1160 to 270 lines.

Refs #21378
